blueprint_v1/patch/deployer/requirements.txt
PyYAML==6.0
kubernetes==24.2.0
Flask==2.2.2
